Avoid violating React invariant when unmounting due to click
If the datepicker/rangepicker was unmounted due to a mouseclick that
happened outside the component, the channel message is received and
handled after React has begun unmounting the component. At this point,
calling `om/get-node` will trigger an "invariant violation" exception.
Oddly enough this all happens before `will-unmount` runs.

The error can be prevented by checking if the owning React component is
mounted before trying to call `get-node` in an async context.