Fix datepicker/rangepicker event listeners leaking
The document-wide click listeners created by datepicker/rangepicker
components were not removed if the component was unmounted. Hence, in a
scenario where the components are mounted and unmounted multiple
times, unused event listeners would leak.

Some bookkeeping is required to hang on to the listener key returned by
`goog.events/listen`, so that we can use it to remove the listener when
unmounting.